% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_adapt.R
\name{compute_adaptation}
\alias{compute_adaptation}
\title{Compute the necessary information for testing adaptation}
\usage{
compute_adaptation(data)
}
\arguments{
\item{data}{A dataframe containing per each study the time series of relevant variables
(i.e. yearly climate values, yearly trait values or yearly selection differentials)
to be analyzed.}
}
\value{
A \var{data.frame}.
}
\description{
\code{compute_adaptation} computes all necessary information, for all
studies, to compare the Weighted Mean Selection (WMS) gradient to the product
of the step 1 and 2.
}
\examples{
data_for_Trait <- prepare_data(data = dat_Trait, temperature = TRUE,
                               phenology = TRUE, morphology = FALSE)
compute_adaptation(data = data_for_Trait[data_for_Trait$id \%in\% c("7"), ])

}
