% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_adapt.R
\name{compute_WMS}
\alias{compute_WMS}
\title{Compute the Weighted Mean Selection (WMS) gradient}
\usage{
compute_WMS(mod_sel)
}
\arguments{
\item{mod_sel}{An object produced by the function \code{\link{fit_cond_id}}
with the settings 'condition = 3'.}
}
\value{
A list with two elements: the Weighted Mean Selection (WMS)
gradient, and the weighted standard error associated with WMS (se).
}
\description{
\code{compute_WMS} computes the WMS gradient based on the fit of step 3 for a
given study. The WMS gradient is produced by predicting the gradient in each
year during the study period of ALL studies (accounting for the realization
of the yearly random effect). The predictions are then averaged accounting
for their uncertainty as measured by the prediction variance of the fit.
}
\note{
The weighted standard error is the square root of a prediction variance.
It represents the prediction uncertainty of a weighted average of yearly
selection gradients, taking into account the prediction covariance of yearly
selection gradients. The weights used for averaging are the inverse of the
yearly prediction variances, to minimize the variance of the average.
}
\examples{
data_for_Trait <- prepare_data(data = dat_Trait, temperature = TRUE,
                               phenology = TRUE, morphology = FALSE)
mod3 <- fit_cond_id(data = data_for_Trait, id = "82", condition = "3")
compute_WMS(mod_sel = mod3)

}
