% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_effects.R
\name{extract_effects}
\alias{extract_effects}
\title{Extract the effect sizes and their standard errors from the model fitted to a single study}
\usage{
extract_effects(list_fit)
}
\arguments{
\item{list_fit}{A fitted object returned by the function \code{\link{fit_cond_id}}.}
}
\value{
A dataframe with the id of the study, the estimated effect sizes and their SE,
the estimated autocorrelation parameter and the results of the LRT. The dataframe also
contains all the meta-data supplied via 'data' argument.
}
\description{
\code{extract_effects} retrieves the effect size and SE for a specified study,
for a given condition.
}
\examples{
dat_clim <- prepare_data(data = dat_Clim, temperature = TRUE,
                         precipitation = FALSE,
                         phenology = TRUE, morphology = TRUE)

fitted_mod <- fit_cond_id(data = dat_clim, id = '30', condition = '1')
test <- extract_effects(fitted_mod)


dat_sel_phen <- prepare_data(data = dat_Sel, temperature = TRUE,
                             precipitation = FALSE,
                             phenology = TRUE, morphology = FALSE)

test_sel_phen <- extract_effects(fit_cond_id(data = dat_sel_phen,
                                 id  = '1', condition = '3'))

dat_phen_prcs <- prepare_data(data = dat_Trait, temperature = TRUE,
                              precipitation = FALSE,
                              phenology = TRUE, morphology = FALSE)
mod_phen_abund <- fit_cond_id(data = dat_phen_prcs, id  = '1', condition = '2b')
test_tr_phen_abund <- extract_effects(mod_phen_abund)

}
