% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_effects_all_ids.R
\name{extract_effects_all_ids}
\alias{extract_effects_all_ids}
\title{Fit model on each study, for a given condition, and extract the effect sizes and their standard errors}
\usage{
extract_effects_all_ids(data, condition, nb_cores)
}
\arguments{
\item{data}{A dataframe containing per each study the time series of relevant variables
(i.e. yearly climate values, yearly trait values or yearly selection differentials)
to be analyzed.}

\item{condition}{A character specifying which condition is to be tested (for more details see
Radchuk et al. (in review)):
\itemize{\item '1' - Condition 1 (effect of year on climate);
\item '2' - Condition 2 (effect of climate on traits);
\item '3' - Condition 3 (testing whether weighted mean selection over years differs from 0);
\item '2b' - Condition 2, but also with abundance included as a predictor;
\item '3b' - testing whether selection changes over years.}}

\item{nb_cores}{An integer indicating how many cores should be used for assessing the LRT by
bootstrap (see spaMM documentation for more details).}
}
\value{
A dataframe containing all effect sizes and SE for the meta-analysis.
}
\description{
\code{extract_effects_all_ids} retrieves the effect size and SE for each study
by calling the function \code{\link{extract_effects}} on each study.
}
\examples{
\dontrun{
nb_cores <- 2L ## increase the number for using more cores
dat_clim <- prepare_data(data = dat_Clim, temperature = TRUE,
                         precipitation = FALSE,
                         phenology = TRUE, morphology = TRUE)
test <- extract_effects_all_ids(data = dat_clim,
                                condition = "1", nb_cores = nb_cores)
head(test)

dat_sel_phen <- prepare_data(data = dat_Sel, temperature = TRUE,
                             precipitation = FALSE,
                             phenology = TRUE, morphology = FALSE)
test_sel <- extract_effects_all_ids(data = dat_sel_phen,
                                    condition = "3", nb_cores = nb_cores)


dat_T_Trait <- prepare_data(data = dat_Trait, temperature = TRUE,
                            precipitation = FALSE,
                            phenology = TRUE, morphology = FALSE)
test_Trait <- extract_effects_all_ids(data = dat_T_Trait,
                                      condition = "2", nb_cores = nb_cores)
}

}
