% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_cond_id}
\alias{fit_cond_id}
\title{Fit a single mixed-effects model to test a given condition for a given study id}
\usage{
fit_cond_id(data, id, condition)
}
\arguments{
\item{data}{A dataframe containing per each study the time series of relevant variables
(i.e. yearly climate values, yearly trait values or yearly selection differentials)
to be analyzed.}

\item{id}{A character specifying the unique study id.}

\item{condition}{A character specifying which condition is to be tested (for more details see
Radchuk et al. (in review)):
\itemize{\item '1' - Condition 1 (effect of year on climate);
\item '2' - Condition 2 (effect of climate on traits);
\item '3' - Condition 3 (testing whether weighted mean selection over years differs from 0);
\item '2b' - Condition 2, but also with abundance included as a predictor;
\item '3b' - testing whether selection changes over years.}}
}
\value{
The list of length 5. The first element is the best full
fitted model returned with the function \code{\link{fit_model}}.
The second element contains results of LRT. The third element
is a condition tested, the forth element is a study id, and the
fifth element is a dataset to what the study belongs (PRCS or PRC).
}
\description{
\code{fit_cond_id} fits a mixed-effects model for a single study id in the
dataset. This model takes into account temporal autocorrelation and includes
the desired fixed-effects predictors.
}
\examples{

dat_trials <- prepare_data(data = dat_Clim,
                           temperature = TRUE, precipitation = FALSE,
                           phenology = TRUE, morphology = TRUE)
test <- fit_cond_id(data = dat_trials, id = '1', condition = '1')

dat_temp_phen <- prepare_data(data = dat_Trait,
                           temperature = TRUE, precipitation = FALSE,
                           phenology = TRUE, morphology = FALSE)
test_trait <- fit_cond_id(data = dat_temp_phen, id = '1', condition = '2')
test_trait$model


}
\seealso{
\code{\link{fit_model}}
}
