% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit the model for a given study id}
\usage{
fit_model(data, formula_full = "~ 1", formula_null = "~ 1",
  resid_formula = "~ 1", prior_weights = NULL, ...)
}
\arguments{
\item{data}{A dataframe containing per each study the time series of relevant variables
(i.e. yearly climate values, yearly trait values or yearly selection differentials)
to be analyzed.}

\item{formula_full}{A character string indicating the formula for the full
model to be fitted.}

\item{formula_null}{A character string indicating the formula for the null
model to be fitted.}

\item{resid_formula}{A character string indicating the formula for the
residual model to be fitted.}

\item{prior_weights}{A numeric vector of prior weights.}

\item{...}{Additional parameters for the function \code{\link[spaMM]{fitme}}.}
}
\value{
A list containing the best full fitted model and the result of
  \code{\link[spaMM]{anova.HLfit}} comparing this model and a null model.
}
\description{
\code{fit_model} fits the model for a given study id. The function first fits
the model using REML with and without the AR1 auto-regressive term. It then
compares the AIC of these two models. The best model is then refitted using ML
and compared to a null model. All models are fitted using the package
\pkg{\link[spaMM]{spaMM}}.
}
\examples{
dat_trials <- prepare_data(data = dat_Clim,
                           temperature = TRUE, precipitation = FALSE,
                           phenology = TRUE, morphology = TRUE)

test <- fit_model(data = dat_trials,
                  formula_full = 'Clim ~ year',
                  formula_null = 'Clim ~ 1')

}
