% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_heterogen.R
\name{get_heterogen}
\alias{get_heterogen}
\title{Get metrics of heterogeneity}
\usage{
get_heterogen(model, digit = 2)
}
\arguments{
\item{model}{A meta-analytical model including studies as random effects.
Heterogeneity metrics are to be computed for this model.}

\item{digit}{An integer indicating how many digits to display on the screen.}
}
\value{
A vector of the calculated heterogeneity metrics:
}
\description{
\code{get_heterogen} computes the heterogeneity metrics for meta-analysis.
}
\details{
This function computes the heterogeneity metrics for the meta-analysis,
including I2, H2, and Q
}
