% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stat_effsizes.R
\name{get_stat_effsizes}
\alias{get_stat_effsizes}
\title{Extracts statistics of interest or effect sizes for a specified list of random-effects meta-analytical models}
\usage{
get_stat_effsizes(list, stat)
}
\arguments{
\item{list}{A list of meta-analytical models from which statistics is to be extracted.}

\item{stat}{A character specifying what should be extracted from the fitted models. Possible are:
LRT' for likelihood-ratio test statistics, 'RandStudy' for the random variance due to the study ID,
'RandPub' for the random variance due to the publication ID, 'globslope' for global effect sizes,
'globse' for the standard errors of global effect sizes.}
}
\value{
A dataframe with three columns: 'chi2_LR', 'df' and 'p_value' if the specified statistics is
'LRT' and a numeric vector of extracted random variances or effect sizes or standard errors otherwise.
}
\description{
Extracts statistics of interest or effect sizes for a specified list of random-effects meta-analytical models
}
\examples{
nb_cores <- min(c(parallel::detectCores(), 10))
mod_T_prcs <- fit_all(data = dat_Clim,
                      temperature = TRUE, precipitation = FALSE,
                      phenology = TRUE, morphology = TRUE,
                      condition = '1', nb_cores = nb_cores,
                      rand_trait = FALSE, fixed = NULL, digit = 3)
mod_Sel_T_phen <- fit_all(data = dat_Sel,
                          temperature = TRUE, precipitation = FALSE,
                          phenology = TRUE, morphology = FALSE,
                          condition = '3', nb_cores = nb_cores,
                          rand_trait = FALSE, fixed = NULL, digit = 3)
mod_list <- list(mod_T_prcs, mod_Sel_T_phen)

test_LRT <- get_stat_effsizes(list = mod_list, stat = 'LRT')
test_LRT1 <- get_stat_effsizes(list = mod_list, stat = 'RandPub')

}
