% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{info_package}
\alias{info_package}
\title{Provide some basic information about the code of the package}
\usage{
info_package()
}
\description{
This is a function for development purposes only and is thus not exported.
This function provides the number of functions and lines of code of the
package.
}
