% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_two_data.R
\name{merge_two_data}
\alias{merge_two_data}
\title{Merge two datasets into a single, PRC (Phenotypic Responses to Climate)
dataset}
\usage{
merge_two_data(dat_nos, dat_prcs, traits = FALSE)
}
\arguments{
\item{dat_nos}{Dataset without selection.}

\item{dat_prcs}{Dataset with selection - PRCS (Phenotypic Responses to
Climate with Selection data).}

\item{traits}{A boolean indicating whether the phenotypic traits are
used as response.}
}
\value{
A dataframe combining the studies for which selection data are
  available with those for which no selection data are available. This
  constitutes a PRC dataset ('Phenotypic Responses to Climate')
}
\description{
\code{merge_two_data} Merges the dataset with selection (PRCS) and the one
that does not have selection data into a single dataset - PRC
}
\details{
This function merges the dataset without selection with the one that
contains selection data, PRCS (Phenotypic Responses to Climate with
Selection data. The unique studies are retained in the resulting PRC
(Phenotypic Responses to Climate) dataset.
}
