% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forest.R
\name{plot_abund_effects}
\alias{plot_abund_effects}
\title{Produce a forest plot, which for each study shows the effect sizes for
 climatic factor, abundance and year}
\usage{
plot_abund_effects(meta_data)
}
\arguments{
\item{meta_data}{A dataframe with rows representing records per studies,
including the effect sizes and required fixed and random effects.}
}
\value{
Plots a forest plot that shows for each study the effect sizes
(and SEs) estimated for the climatic variable, abundance and year.
}
\description{
Produce a forest plot, which for each study shows the effect sizes for
 climatic factor, abundance and year
}
