% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_supplement.R
\name{plot_eff_dur_firstY}
\alias{plot_eff_dur_firstY}
\title{Plots slopes of temperature vs. the duration of the studies
and vs. the first year in time series}
\usage{
plot_eff_dur_firstY(data, meta_obj)
}
\arguments{
\item{data}{A dataframe containing per each study the time series of relevant variables
(i.e. yearly climate values, yearly trait values or yearly selection differentials)
to be analyzed.}

\item{meta_obj}{A fitted meta-analytical model object.}
}
\value{
Plots three plots: 1. the effect of year on temperature vs
the number of years in the time series; 2. the effect of year on
temperature vs the first year in the time series; and 3. the first
year in the time series vs the number of years.
}
\description{
Plots slopes of temperature vs. the duration of the studies
and vs. the first year in time series
}
\examples{
 \dontrun{
     mod_T_prc <- fit_all(data = dat_Clim_prc,
                          temperature = TRUE, precipitation = FALSE,
                          phenology = TRUE, morphology = TRUE,
                          condition = '1', nb_cores = nb_cores,
                          rand_trait = FALSE, fixed = NULL, digit = digit)
     plot_eff_dur_firstY(data = dat_Clim_prc, meta_obj = mod_T_prc)
 }
}
