% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forest.R
\name{plot_forest}
\alias{plot_forest}
\title{Produce a forest plot with effect sizes for each study id in PRCS dataset and global
effects for both PRCS and PRC datasets}
\usage{
plot_forest(meta_obj1, meta_obj2 = NULL, list_extra_meta_obj = NULL,
  sort = c("Species", "Study_Authors", "Fitness_Categ"),
  increasing = TRUE, labels = c(traits = TRUE, fitness = TRUE, country
  = TRUE, authors = TRUE), pdf_basename = NULL, mar = c(4, 10, 2, 2))
}
\arguments{
\item{meta_obj1}{A fitted meta-analytical model object whose global slopes and per-study effect
sizes will be displayed on the plot.}

\item{meta_obj2}{An (optional) second fitted meta-analytical model object whose
global slopes and per-study effect sizes will be displayed on the plot.}

\item{list_extra_meta_obj}{A list of meta-analytical model objects to be used for displaying the
global effects (either across all studies or in response to a fixed predictor).}

\item{sort}{A vector specifying the variables to use for sorting the labels.}

\item{increasing}{A Boolean specifying whether the slopes should be sorted
in increasing (TRUE) or decreasing (FALSE) order.}

\item{labels}{A vector specifying what should be used as labels for each study.
See Details for kinds of labels available.}

\item{pdf_basename}{A character specifying the name of the .pdf for
the produced plot.}

\item{mar}{A vector specifying the plot margins, analogously to \code{\link[graphics]{par}}.}
}
\value{
Plots a forest plot with effect sizes (and SEs) for each study,
and global effect size(s) on the bottom. The data used for the plot are
returned (invisibly).
}
\description{
\code{plot_forest} Plots a forest plot displaying the effect sizes
for each study and global effects across the studies
}
\details{
This function first fits separate mixed-effects models per study id. All the
models include predictors of interest. The possible types of labels are:
\itemize{
\item traits - the specific trait reported for each study;
\item fitness - the fitness measure used in each study;
\item country - a 2-letter country code;
\item authors - if there are more than two authors, the name of
the first one, otherwise the name of the first two.} If the label is set to
TRUE (default), it will be displayed on the plot.
}
\examples{
nb_cores <- min(c(parallel::detectCores(), 10))
meta_Trait_phen <- fit_all(data = dat_Trait, temperature = TRUE,
                           precipitation = FALSE, phenology = TRUE,
                           morphology = FALSE, condition = '2',
                           nb_cores = nb_cores, rand_trait = FALSE,
                           fixed = NULL, digit = 3)

meta_Trait_morph <- fit_all(data = dat_Trait, temperature = TRUE,
                            precipitation = FALSE, phenology = FALSE,
                            morphology = TRUE, condition = '2',
                            nb_cores = nb_cores, rand_trait = FALSE,
                            fixed = NULL, digit = 3)

meta_Trait_phen_Tax <- fit_all(data = dat_Trait, temperature = TRUE,
                               precipitation = FALSE, phenology = TRUE,
                               morphology = FALSE, condition = '2',
                               nb_cores = nb_cores, rand_trait = FALSE,
                               fixed = 'Taxon', digit = 3)

test <- plot_forest(meta_obj1 = meta_Trait_phen$meta_res,
                    meta_obj2 = meta_Trait_morph$meta_res,
                    list_extra_meta_obj = list(meta_Trait_phen_Tax$meta_res),
                    sort = c("Species", "Study_Authors", "Fitness_Categ"),
                    increasing = TRUE,
                    labels = c(traits = TRUE,
                               fitness = TRUE,
                               country = TRUE,
                               authors = TRUE),
                    pdf_basename = 'test')

}
