% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_per_id.R
\name{plot_per_id}
\alias{plot_per_id}
\title{Plot data and fitted model per each study id}
\usage{
plot_per_id(data, temperature = FALSE, precipitation = FALSE,
  phenology = FALSE, morphology = FALSE, id, condition)
}
\arguments{
\item{data}{A dataframe containing per each study the time series of relevant variables
(i.e. yearly climate values, yearly trait values or yearly selection differentials)
to be analyzed.}

\item{temperature}{A boolean indicating if the temperature data should be extracted (default is FALSE).}

\item{precipitation}{A boolean indicating if the precipitation data should be extracted (default is FALSE).}

\item{phenology}{A boolean indicating if the phenological data should be extracted (default is FALSE).}

\item{morphology}{A boolean indicating if the morphological data should be extracted (default is FALSE).}

\item{id}{A character specifying the unique study id.}

\item{condition}{A character specifying which condition is to be tested (for more details see
Radchuk et al. (in review)):
\itemize{\item '1' - Condition 1 (effect of year on climate);
\item '2' - Condition 2 (effect of climate on traits);
\item '3' - Condition 3 (testing whether weighted mean selection over years differs from 0);
\item '2b' - Condition 2, but also with abundance included as a predictor;
\item '3b' - testing whether selection changes over years.}}
}
\value{
A plot showing the raw data points and a fitted line.
}
\description{
\code{plot_per_id} Plots raw data and a fitted model for a single study id in the dataset.
This model is fitted with \code{\link{fit_cond_id}}. Plot also visualizes the
p value for the fit of the model.
}
\details{
This function fits a single model per study id. This model includes
predictors of interest and accounts for temporal autocorrelation. Predictors
of interest can be: \itemize{ \item a year (if assessing condition 1, change in climatic
factor), \item a climatic factor and a year (if assessing condition 2, change in
the trait with climate).}
}
\examples{
plot_per_id(data = dat_Clim, temperature = TRUE,
            precipitation = FALSE, phenology = TRUE,
            morphology = TRUE, id = '1', condition = '1')

plot_per_id(data = dat_Trait, temperature = TRUE,
            precipitation = FALSE, phenology = TRUE,
            morphology = FALSE, id = '1', condition = '2')

}
