% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raw.R
\name{plot_raw}
\alias{plot_raw}
\title{Plot raw data and the fitted model for the selected set of studies}
\usage{
plot_raw(data, temperature = FALSE, precipitation = FALSE,
  phenology = FALSE, morphology = FALSE, condition, id_to_do)
}
\arguments{
\item{data}{A dataframe containing per each study the time series of relevant variables
(i.e. yearly climate values, yearly trait values or yearly selection differentials)
to be analyzed.}

\item{temperature}{A boolean indicating if the temperature data should be extracted (default is FALSE).}

\item{precipitation}{A boolean indicating if the precipitation data should be extracted (default is FALSE).}

\item{phenology}{A boolean indicating if the phenological data should be extracted (default is FALSE).}

\item{morphology}{A boolean indicating if the morphological data should be extracted (default is FALSE).}

\item{condition}{A character specifying which condition is to be tested (for more details see
Radchuk et al. (in review)):
\itemize{\item '1' - Condition 1 (effect of year on climate);
\item '2' - Condition 2 (effect of climate on traits);
\item '3' - Condition 3 (testing whether weighted mean selection over years differs from 0);
\item '2b' - Condition 2, but also with abundance included as a predictor;
\item '3b' - testing whether selection changes over years.}}

\item{id_to_do}{A list of study ids for which to plot the raw data and fitted models.}
}
\value{
A plot showing the raw data points and fitted models for a
specified subset of studies.
}
\description{
\code{plot_raw} Plots raw data and fitted models for a
specified set of studies.
}
\examples{
\dontrun{
pdf('./test_sel.pdf', height = 10, width = 7)
par(oma = c(2, 2, 0, 0), mar = c(1, 2, 3, 1))
plot_raw(data = dat_Sel, temperature = TRUE,
         precipitation = FALSE, phenology = TRUE,
         morphology = FALSE, condition = '3',
         id_to_do = c(1:19))
dev.off()

pdf('./test_traits.pdf', height = 10, width = 7)
par(oma = c(2, 2, 0, 0), mar = c(1, 2, 3, 1))
plot_raw(data = dat_Trait, temperature = TRUE,
         precipitation = FALSE, phenology = TRUE,
         morphology = FALSE, condition = '2',
         id_to_do = c(1:20))
dev.off()
}

}
