% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sel_per_id.R
\name{plot_sel_per_id}
\alias{plot_sel_per_id}
\title{Plot selection data and WMS per each study id}
\usage{
plot_sel_per_id(data, temperature = FALSE, precipitation = FALSE,
  phenology = FALSE, morphology = FALSE, id, condition)
}
\arguments{
\item{data}{A dataframe containing per each study the time series of relevant variables
(i.e. yearly climate values, yearly trait values or yearly selection differentials)
to be analyzed.}

\item{temperature}{A boolean indicating if the temperature data should be extracted (default is FALSE).}

\item{precipitation}{A boolean indicating if the precipitation data should be extracted (default is FALSE).}

\item{phenology}{A boolean indicating if the phenological data should be extracted (default is FALSE).}

\item{morphology}{A boolean indicating if the morphological data should be extracted (default is FALSE).}

\item{id}{A character specifying the unique study id.}

\item{condition}{A character specifying which condition is to be tested (for more details see
Radchuk et al. (in review)):
\itemize{\item '1' - Condition 1 (effect of year on climate);
\item '2' - Condition 2 (effect of climate on traits);
\item '3' - Condition 3 (testing whether weighted mean selection over years differs from 0);
\item '2b' - Condition 2, but also with abundance included as a predictor;
\item '3b' - testing whether selection changes over years.}}
}
\value{
A plot showing the raw selection data and an estimate of WMS.
}
\description{
\code{plot_sel_per_id} Plots raw selection data and an estimate of WMS for a single study id in the dataset.
Plot also visualizes the p value for the fit of the model.
}
\examples{
plot_sel_per_id(data = dat_Sel, temperature = TRUE,
                precipitation = FALSE, phenology = TRUE,
                morphology = FALSE, id = '1', condition = '3')

}
