% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_forestplot.R
\name{prepare_colours}
\alias{prepare_colours}
\title{Create the colours for the forest plots}
\usage{
prepare_colours(slopes_obj)
}
\arguments{
\item{slopes_obj}{A list returned by \code{\link{prepare_slopes}} that contains
the lists with values for the slopes, their confidence intervals and labels.}
}
\value{
The list provided as an input with the colours added as a new row of
each data frame contained in this list.
}
\description{
\code{prepare_colours} prepares the colours for the forest plots. This is
an internal function used by \code{\link{plot_forest}}.
}
\examples{
\dontrun{
nb_cores <- 2L
meta_Sel_phen <- fit_all(data = dat_Sel,
                         temperature = TRUE, precipitation = FALSE,
                         phenology = TRUE, morphology = FALSE,
                         condition = '3', nb_cores = nb_cores,
                         rand_trait = FALSE, fixed = NULL, digit = 3)

meta_Sel_morph <- fit_all(data = dat_Sel,
                          temperature = TRUE, precipitation = FALSE,
                          phenology = FALSE, morphology = TRUE,
                          condition = '3', nb_cores = nb_cores,
                          rand_trait = FALSE, fixed = NULL, digit = 3)


test1 <- prepare_slopes(meta_obj1 = meta_Sel_phen$meta_res)

test2 <- prepare_slopes(meta_obj2 = meta_Sel_phen$meta_res)

test3 <- prepare_slopes(meta_obj1 = meta_Sel_phen$meta_res,
                        meta_obj2 = meta_Sel_morph$meta_res)

test_col_1 <- prepare_colours(test1)
test_col_2 <- prepare_colours(test2)
test_col_3 <- prepare_colours(test3)
}

}
\seealso{
\code{\link{plot_forest}}.
}
