% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Subsets the dataset according to the climatic variable and phenotypic category}
\usage{
prepare_data(data, temperature = FALSE, precipitation = FALSE,
  phenology = FALSE, morphology = FALSE)
}
\arguments{
\item{data}{A dataframe containing per each study the time series of relevant variables
(i.e. yearly climate values, yearly trait values or yearly selection differentials)
to be analyzed.}

\item{temperature}{A boolean indicating if the temperature data should be extracted (default is FALSE).}

\item{precipitation}{A boolean indicating if the precipitation data should be extracted (default is FALSE).}

\item{phenology}{A boolean indicating if the phenological data should be extracted (default is FALSE).}

\item{morphology}{A boolean indicating if the morphological data should be extracted (default is FALSE).}
}
\value{
A dataframe that is a subset of the dataframe supplied via
the argument 'data'. This dataframe is subset according to the specified
requirements.
}
\description{
\code{prepare_data} subsets the dataset according to the specified conditions.
}
