% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_forestplot.R
\name{prepare_data_plot}
\alias{prepare_data_plot}
\title{Prepare the data for the forest plots}
\usage{
prepare_data_plot(slopes_obj)
}
\arguments{
\item{slopes_obj}{A list returned by \code{\link{prepare_slopes}} that contains
the lists with values for the slopes, their confidence intervals and labels.}
}
\value{
A dataframe with the information used for plotting.
}
\description{
\code{prepare_data_plot} prepares the data for the forest plots. This is
an internal function used by \code{\link{plot_forest}}.
}
\examples{
\dontrun{
nb_cores <- 2L
meta_Sel_phen <- fit_all(data = dat_Sel,
                         temperature = TRUE, precipitation = FALSE,
                         phenology = TRUE, morphology = FALSE,
                         condition = '3', nb_cores = nb_cores,
                         rand_trait = FALSE, fixed = NULL, digit = 3)

meta_Sel_morph <- fit_all(data = dat_Sel,
                          temperature = TRUE, precipitation = FALSE,
                          phenology = FALSE, morphology = TRUE,
                          condition = '3', nb_cores = nb_cores,
                          rand_trait = FALSE, fixed = NULL, digit = 3)

lab1 <- prepare_labels(meta_obj = meta_Sel_phen$meta_res,
                       labels = c(traits = TRUE,
                                  fitness = TRUE,
                                  country = TRUE,
                                  authors = TRUE),
                      sort = c("Species"),
                      increasing = TRUE)

lab2 <-  prepare_labels(meta_obj = meta_Sel_morph$meta_res,
                       labels = c(traits = TRUE,
                                  fitness = TRUE,
                                  country = TRUE,
                                  authors = TRUE),
                      sort = c("Species"),
                      increasing = TRUE)


test1 <- prepare_slopes(meta_obj1 = meta_Sel_phen$meta_res,
                       labels_obj1 = lab1,
                       labels_obj2 = lab2)

test2 <- prepare_slopes(meta_obj1 = meta_Sel_phen$meta_res,
                        meta_obj2 = meta_Sel_morph$meta_res,
                        labels_obj1 = lab1,
                        labels_obj2 = lab2,
                        list_extra_meta_obj = list(meta_Sel_phen$meta_res,
                                                  meta_Sel_morph$meta_res,
                                                  mod_Sel_T_phen_Fitn$meta_res))

test_col_1 <- prepare_colours(test1)
test_col_2 <- prepare_colours(test2)
prep_data_test1 <- prepare_data_plot(test_col_1)
prep_data_test2 <- prepare_data_plot(test_col_2)
}

}
\seealso{
\code{\link{plot_forest}}.
}
