% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_forestplot.R
\name{prepare_labels}
\alias{prepare_labels}
\title{Create the labels for the forest plots}
\usage{
prepare_labels(meta_obj = NULL, labels, sort, increasing)
}
\arguments{
\item{meta_obj}{A fitted meta-analytical model object.}

\item{labels}{A vector specifying what should be used as labels for each study.
See Details for kinds of labels available.}

\item{sort}{A vector specifying the variables to use for sorting the labels.}

\item{increasing}{A Boolean specifying whether the slopes should be sorted
in increasing (TRUE) or decreasing (FALSE) order.}
}
\value{
A list with two elements: the labels and the order of the study.

The possible types of labels are:
\itemize{
\item traits - the specific trait reported for each study;
\item fitness - the fitness measure used in each study;
\item country - a 2-letter country code;
\item authors - if there are more than two authors, the name of
the first one, otherwise the name of the first two.} If the label is set to
TRUE (default), it will be displayed on the plot.
}
\description{
\code{prepare_labels} prepares the labels for the forest plots. This is
an internal function used by \code{\link{plot_forest}}.
}
\examples{
\dontrun{
dat_sel_phen <- prepare_data(data = dat_Sel,
                             temperature = TRUE, precipitation = FALSE,
                             phenology = TRUE, morphology = FALSE)
test_sel <- extract_effects_all_ids(data = dat_sel_phen,
                                    condition = '3', nb_cores = 4L)
test_meta_sel <- fit_meta(test_sel)
prepare_labels(meta_obj = test_meta_sel,
               labels = c(traits = TRUE,
                          fitness = TRUE,
                          country = TRUE,
                          authors = TRUE),
               sort = c("Species", "Study_Authors", "Fitness_Categ"),
               increasing = TRUE)
}


}
\seealso{
\code{\link{plot_forest}}
}
