% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_forestplot.R
\name{prepare_slopes}
\alias{prepare_slopes}
\title{Prepare the slopes for the forest plots}
\usage{
prepare_slopes(meta_obj1 = NULL, meta_obj2 = NULL,
  list_extra_meta_obj = NULL, labels_obj1 = NULL, labels_obj2 = NULL)
}
\arguments{
\item{meta_obj1}{A fitted meta-analytical model object whose global slopes and per-study effect
sizes will be displayed on the plot.}

\item{meta_obj2}{An (optional) second fitted meta-analytical model object whose
global slopes and per-study effect sizes will be displayed on the plot.}

\item{list_extra_meta_obj}{A list of meta-analytical model objects to be used for displaying the
global effects (either across all studies or in response to a fixed predictor).}

\item{labels_obj1}{A list of labels to be used for slopes extracted from
meta_obj1 object. This list is returned by the function \code{\link{prepare_labels}}.}

\item{labels_obj2}{A list of labels to be used for slopes extracted from
meta_obj2 object. This list is returned by the function \code{\link{prepare_labels}}}
}
\value{
A list containing the slopes and other information required
for the forest plot.
}
\description{
\code{prepare_slopes} extracts the slopes needed for the forest plots from
the fitted model objects.
}
\examples{
\dontrun{
nb_cores <- min(c(parallel::detectCores(), 10))
meta_Sel_phen <- fit_all(data = dat_Sel,
                         temperature = TRUE, precipitation = FALSE,
                         phenology = TRUE, morphology = FALSE,
                         condition = '3', nb_cores = nb_cores,
                         rand_trait = FALSE, fixed = NULL, digit = 3)

meta_Sel_morph <- fit_all(data = dat_Sel,
                          temperature = TRUE, precipitation = FALSE,
                          phenology = FALSE, morphology = TRUE,
                          condition = '3', nb_cores = nb_cores,
                          rand_trait = FALSE, fixed = NULL, digit = 3)

meta_Sel_phen_Fitn <- fit_all(data = dat_Sel,
                              temperature = TRUE, precipitation = FALSE,
                              phenology = TRUE, morphology = FALSE,
                              condition = '3', nb_cores = nb_cores,
                              rand_trait = FALSE, fixed = 'Fitness_Categ',
                              digit = 3)

test1 <- prepare_slopes(meta_obj1 = meta_Sel_phen$meta_res)

test2 <- prepare_slopes(meta_obj2 = meta_Sel_phen$meta_res)

test3 <- prepare_slopes(meta_obj1 = meta_Sel_phen$meta_res,
                        meta_obj2 = meta_Sel_morph$meta_res)

test4 <- prepare_slopes(meta_obj1 = meta_Sel_phen$meta_res,
                        list_extra_meta_obj =
                        list(meta_Sel_phen_Fitn$meta_res))

test5 <- prepare_slopes(meta_obj1 = meta_Sel_phen$meta_res,
                        meta_obj2 = meta_Sel_morph$meta_res,
                        list_extra_meta_obj =
                        list(meta_Sel_phen_Fitn$meta_res))

test6 <- prepare_slopes(meta_obj1 = meta_Sel_phen_Fitn$meta_res,
                        list_extra_meta_obj =
                        list(meta_Sel_morph$meta_res,
                        meta_Sel_phen$meta_res))
}


}
\seealso{
\code{\link{plot_forest}}
}
