% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_forestplot.R
\name{prepare_xlabs}
\alias{prepare_xlabs}
\title{Create the x-axis labels for the forest plots}
\usage{
prepare_xlabs(meta_obj)
}
\arguments{
\item{meta_obj}{A fitted meta-analytical model object.}
}
\value{
A vector of length one with the x-axis label. If \code{meta_obj} is
NULL, the function returns NULL.
}
\description{
\code{prepare_xlabs} prepares the x-axis label for the forest plots. This is
an internal function used by \code{\link{plot_forest}}.
}
\seealso{
\code{\link{plot_forest}}.
}
