% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_functions.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{binom.confint}
\title{Export functions}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{binom}{\code{\link[binom]{binom.confint}}}
}}

