% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_xlsx.R
\name{save_xlsx}
\alias{save_xlsx}
\title{Save a table in the *.xlsx format}
\usage{
save_xlsx(table, table_basename, first_column_name = NULL, typeTab)
}
\arguments{
\item{table}{The dataframe to be exported as *.xlsx}

\item{table_basename}{A string indicating the base name for the file to be created.}

\item{first_column_name}{A string indicating the name of the first column which is taken from the row names of the table}

\item{typeTab}{A character specifying what type of table to output: 'LRT' for a table
with likelihood-ratio statistics and random variances, 'efSizes' for a table with
estimated effect sizes and their standard errors, and 'heterog' for a table
of heterogeneity metrics.}
}
\value{
The table (but the return is invisible).
}
\description{
This function is called internally by other functions. It is a wrapper for
the function \code{\link[openxlsx:write.xlsx]{write.xlsx}} of the package \pkg{openxlsx}.
}
\examples{
d <- data.frame(test = 1:5, id = LETTERS[1:5],
                a= seq(from = 0.01, to = 0.05, by = 0.01),
                test = 1:5, id = LETTERS[1:5],
                p = c('<0.001', '0.002', '0.260', '<0.001', '0.987'))
dir.create('tables')
save_xlsx(table = d, table_basename = './tables/test1',
          typeTab = 'LRT')

}
