% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_supplement.R
\name{tab_efSizes_ST1}
\alias{tab_efSizes_ST1}
\title{Produces a table with summary statistics for the specified set
of models, as in Supplementary Table S1}
\usage{
tab_efSizes_ST1(model_list, table_basename)
}
\arguments{
\item{model_list}{A list of the models from which to extract the summary statistics for the table.}

\item{table_basename}{A string indicating the file name for the exported table}
}
\value{
Saves a table analogous to the Supplementary Table S1 to the specified location.
}
\description{
Produces a table with summary statistics for the specified set
of models, as in Supplementary Table S1
}
