% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_supplement.R
\name{tab_efSizes_ST4}
\alias{tab_efSizes_ST4}
\title{Produces a table with summary statistics for the models conducted
for sensitivity analysis, as in Supplementary Table S4}
\usage{
tab_efSizes_ST4(model_list)
}
\arguments{
\item{model_list}{A list of the models from which to extract the summary statistics for the table.}
}
\value{
Returns a table analogous to the Supplementary Table S4.
}
\description{
Produces a table with summary statistics for the models conducted
for sensitivity analysis, as in Supplementary Table S4
}
