% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_supplement.R
\name{tab_heterog}
\alias{tab_heterog}
\title{Produces a table with the heterogeneity metrics for the specified list of models,
this table is analogous to the Supplementary Table S7.}
\usage{
tab_heterog(model_list, table_basename)
}
\arguments{
\item{model_list}{A list of the models from which to extract the summary statistics for the table.}

\item{table_basename}{A string indicating the file name for the exported table}
}
\value{
Saves a table analogous to the Supplementary Table S7 to the specified location.
}
\description{
Produces a table with the heterogeneity metrics for the specified list of models,
this table is analogous to the Supplementary Table S7.
}
