% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_supplement.R
\name{tab_LRT_ST5}
\alias{tab_LRT_ST5}
\title{Produces a table with LRT statistics and variation due to random effects,
for the models conducted for sensitivity analysis, as in Supplementary Table S5}
\usage{
tab_LRT_ST5(model_list, table_basename)
}
\arguments{
\item{model_list}{A list of the models from which to extract the summary statistics for the table.}

\item{table_basename}{A string indicating the file name for the exported table}
}
\value{
Saves a table analogous to the Supplementary Table S5 to the specified location.
}
\description{
Produces a table with LRT statistics and variation due to random effects,
for the models conducted for sensitivity analysis, as in Supplementary Table S5
}
